Move CssClass3 away from NodeUpdateImports to avoid side effects to tests
