+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/scanner/ScannerTestComponents.java
@@ -281,14 +281,6 @@ public class ScannerTestComponents {
     public static class CssClass2 extends CssClass1 {
     }
 
-    @Route("css-route-3")
-    @CssImport("./foo.css")
-    @CssImport(value = "./bar.css")
-    @CssImport(value = "./foofoo.css")
-    @CssImport(value = "./foobar.css")
-    public static class CssClass3 {
-    }
-
     @JsModule("dynamic-component.js")
     public static class DynamicComponentClass extends Component {
     }
